+++ src/resources/views/forms/checkbox.blade.php
@@ -1,0 +1,3 @@
+<div class="form-group checkbox-inline">
+	<label>{!! Form::checkbox($name, 1, ($value == '1')) !!} {{ $label }}</label>
+</div>