+++ src/resources/views/forms/select.blade.php
@@ -1,4 +1,4 @@
 <div class="form-group">
 	<label>{{ $label }}</label>
-	{!! Form::select($name, $options, null, ['class' => 'form-control']) !!}
+	{!! Form::select($name, $options, (isset($value)) ? $value : null, ['class' => 'form-control']) !!}
 </div>